Uploaded mntnas script
This script is used to mount a locally hosted NAS to a subdirectory in whoever runs it's home directory. It assumes that you're wanting to mount to /home/$USER/nas and that the mount you're attempting to mount is under the same username you are currently logged in as. It also assumes you have a .creds file located in your home directory and gives you a suggestion to chmod 600 this file so your user is the only one able to read it. If one doesn't exist it gives details as to what you would need to do in order to resolve this.

It also creates the /home/$USER/nas directory if it doesn't exist before mounting.

The variable nasip can be changed to whatever IP or FQDN you need it to be for your server.